+++ plugins/fabric.zsh
@@ -3,7 +3,7 @@ exists fabric || {
   install-fabric() {
     info "Installing fabric..."
     _run_remote_installer "https://raw.githubusercontent.com/danielmiessler/fabric/main/scripts/installer/install.sh" "bash" \
-      --env "INSTALL_DIR=$CUSTOM_TOOLS_DIR"
+      --env "INSTALL_DIR=$CUSTOM_TOOLS_DIR" || return
     info "Run 'fabric --setup' to configure API keys"
     reload
   }
@@ -26,16 +26,16 @@ _fabric_load_patterns() {
 _fabric_load_patterns
 
 yt() {
-  if (( $# == 0 || $# > 2 )); then
+  local transcript_flag="--transcript"
+  if [[ ${1-} == "-t" || ${1-} == "--timestamps" ]]; then
+    transcript_flag="--transcript-with-timestamps"
+    shift
+  fi
+
+  if (( $# != 1 )); then
     builtin print -r -- "Usage: yt [-t | --timestamps] youtube-link"
     builtin print -r -- "Use the '-t' flag to get the transcript with timestamps."
     return 1
-  fi
-
-  local transcript_flag="--transcript"
-  if [[ $1 == "-t" || $1 == "--timestamps" ]]; then
-    transcript_flag="--transcript-with-timestamps"
-    shift
   fi
 
   local video_link="$1"
@@ -44,7 +44,7 @@ yt() {
 
 uninstall-fabric() {
   info "Uninstalling fabric..."
-  command rm -f -- "${commands[fabric]}"
+  command rm -f -- "${commands[fabric]}" || return
   command rm -rf -- "$_fabric_config_dir"
   reload
 }
@@ -56,7 +56,7 @@ _update_fabric() {
 }
 
 update-fabric() {
-  _update_fabric
+  _update_fabric || return
   reload
 }
