+++ plugins/wezterm/wezterm.plugin.zsh
@@ -18,7 +18,7 @@ if exists wezterm; then
   if exists brew; then
     uninstall-wezterm() {
       info "Uninstalling wezterm..."
-      command brew uninstall --cask wezterm
+      command brew uninstall --cask wezterm || return
       command rm -rf -- "$HOME/.config/wezterm"
       reload
     }
@@ -26,7 +26,7 @@ if exists wezterm; then
 elif exists brew; then
   install-wezterm() {
     info "Installing wezterm..."
-    command brew install --no-ask --cask wezterm
+    command brew install --no-ask --cask wezterm || return
     _wezterm_restore_config
     reload
   }
